``` 0.001277042 0.0007755817 0.0007465696 0.001284464 0.0007965405 0.0006198856 0.001249023 0.0007745337 The time was 00:00:08.0141063 for: iteration 0 Training: 0.975654058159393 Testing: 0.143756392116372(9568/9984) ```
Yeah, it works. I can see that LRs can dip below 0 at certain points. Maybe a LR of -13 is too aggressive when combined with KFAC.

Let me play around with it for a bit.

```
{rate = {covariance = 0.85; minibatch_size = 128; weight = 0.0001220703}}
Starting timing for: iteration 0
The time was 00:00:07.6880764 for: iteration 0
Training: 0.975819460408858
Testing: 0.144104263255707(9565/9984)
Starting timing for: iteration 1
The time was 00:00:07.6526792 for: iteration 1
Training: 0.0876943330458588
Testing: 0.089848726982615(9714/9984)
Starting timing for: iteration 2
The time was 00:00:07.6414670 for: iteration 2
Training: 0.0339493323388633
Testing: 0.0846568644045851(9757/9984)
Starting timing for: iteration 3
The time was 00:00:07.6304338 for: iteration 3
Training: 0.00920315818202808
Testing: 0.0726491360783024(9796/9984)
Starting timing for: iteration 4
The time was 00:00:07.6238956 for: iteration 4
Training: 0.00269578546179288
Testing: 0.0697441918681855(9800/9984)
```

Let me try lowering the weight rate for a bit.

```
{rate = {covariance = 0.85; minibatch_size = 128; weight = 6.103516E-05}}
Starting timing for: iteration 0
The time was 00:00:07.6825819 for: iteration 0
Training: 1.30133077228127
Testing: 0.136047276781681(9591/9984)
Starting timing for: iteration 1
The time was 00:00:07.6072374 for: iteration 1
Training: 0.101763807708033
Testing: 0.0959213923010677(9687/9984)
Starting timing for: iteration 2
The time was 00:00:07.6314119 for: iteration 2
Training: 0.0311572491504753
Testing: 0.0785302132920687(9759/9984)
Starting timing for: iteration 3
The time was 00:00:07.6143260 for: iteration 3
Training: 0.00879925334680966
Testing: 0.0646379321589894(9800/9984)
Starting timing for: iteration 4
The time was 00:00:07.6436666 for: iteration 4
Training: 0.00439325660197502
Testing: 0.0611364875209498(9800/9984)
```

Hmmm, let me try increasing it.

```
{rate = {covariance = 0.85; minibatch_size = 128; weight = 0.0004882813}}
Starting timing for: iteration 0
The time was 00:00:07.7205558 for: iteration 0
Training: 0.825106543505498
Testing: 0.39370183503697(9320/9984)
Starting timing for: iteration 1
The time was 00:00:07.6695028 for: iteration 1
Training: 0.302760790818959
Testing: 0.253871187433759(9559/9984)
Starting timing for: iteration 2
The time was 00:00:07.7614337 for: iteration 2
Training: 0.114540454111426
Testing: 0.165279100670957(9691/9984)
Starting timing for: iteration 3
The time was 00:00:07.6890662 for: iteration 3
Training: 0.0613634389279531
Testing: 0.19695235989557(9657/9984)
Starting timing for: iteration 4
The time was 00:00:07.6974159 for: iteration 4
Training: 0.0460337604950821
Testing: 0.144203358368171(9714/9984)
```

This is with -11. Let me try -12.

```
{rate = {covariance = 0.85; minibatch_size = 128; weight = 0.0002441406}}
Starting timing for: iteration 0
The time was 00:00:07.7430963 for: iteration 0
Training: 0.783311469390845
Testing: 0.228146003102526(9428/9984)
Starting timing for: iteration 1
The time was 00:00:07.7024624 for: iteration 1
Training: 0.153601316387644
Testing: 0.129115197617349(9674/9984)
Starting timing for: iteration 2
The time was 00:00:07.7602653 for: iteration 2
Training: 0.0398877249112685
Testing: 0.1050373349941(9731/9984)
Starting timing for: iteration 3
The time was 00:00:07.7233910 for: iteration 3
Training: 0.0103518445158
Testing: 0.0870191330075333(9785/9984)
Starting timing for: iteration 4
The time was 00:00:07.6949915 for: iteration 4
Training: 0.00205087877513786
Testing: 0.0811956085505629(9782/9984)
```

It seems -13 or -14 works best. Let me try -15 next.

```
{rate = {covariance = 0.85; minibatch_size = 128; weight = 3.051758E-05}}
Starting timing for: iteration 0
The time was 00:00:07.6749483 for: iteration 0
Training: 1.87817092287816
Testing: 0.210282895952845(9366/9984)
Starting timing for: iteration 1
The time was 00:00:07.6786353 for: iteration 1
Training: 0.140644627159589
Testing: 0.0976882457758527(9679/9984)
Starting timing for: iteration 2
The time was 00:00:07.6546315 for: iteration 2
Training: 0.0432088367447444
Testing: 0.0759579567190182(9755/9984)
Starting timing for: iteration 3
The time was 00:00:07.6159434 for: iteration 3
Training: 0.0128824701915176
Testing: 0.060238286310871(9804/9984)
Starting timing for: iteration 4
The time was 00:00:07.6062353 for: iteration 4
Training: 0.00359972161095357
Testing: 0.0602788858851478(9815/9984)
```

I already tried -14, but let me do so again. As expected it is hard to expect top notch performance on the first iteration unless the LR is especially well tuned.

```
{rate = {covariance = 0.85; minibatch_size = 128; weight = 6.103516E-05}}
Starting timing for: iteration 0
The time was 00:00:07.6933618 for: iteration 0
Training: 1.30133077228127
Testing: 0.136047276781681(9591/9984)
Starting timing for: iteration 1
The time was 00:00:07.6285612 for: iteration 1
Training: 0.101764185897883
Testing: 0.0959351275766191(9688/9984)
Starting timing for: iteration 2
The time was 00:00:07.6494353 for: iteration 2
Training: 0.0311881141955631
Testing: 0.0788113110784952(9753/9984)
Starting timing for: iteration 3
The time was 00:00:07.6240089 for: iteration 3
Training: 0.00908208965953453
Testing: 0.0635036488880966(9801/9984)
Starting timing for: iteration 4
The time was 00:00:07.6225195 for: iteration 4
Training: 0.00445235936866892
Testing: 0.0617092425988327(9800/9984)
```

I guess I'll make -14 my default. Before I move on, I need to test the multiplicative updates more. I want to try higher settings than -7.