``` The learning rate is 2 ** -14 Starting timing for: iteration 0 1E-06 1E-06 1E-06 1E-06 68.84859 -17.42059 The time was 00:00:00.1450908 for: iteration 0 Training: NaN Training diverged. Aborting... ```
I am at a loss. Should I divide by k instead of minibatch size then? Let me give it a shot.

```
{rate = {covariance = 0.0002616634; minibatch_size = 64; weight = 6.103516E-05}}
The learning rate is 2 ** -14
Starting timing for: iteration 0
1E-06
1E-06
1E-06
1E-06
0.01680974
-0.004252076
The time was 00:00:00.1423394 for: iteration 0
Training: NaN
Training diverged. Aborting...
```

This escalated quickly. Let me try turning off KFAC.

```
The learning rate is 2 ** -14
Starting timing for: iteration 0
1E-06
1E-06
1E-06
1E-06
-1.012032E-05
1.464304E-05
-6.138935E-05
3.974701E-05
-6.102521E-05
8.112727E-05
-2.100929E-05
0.000152341
5.107901E-05
0.0001944204
4.377289E-05
0.0001703357
0.0003345697
-8.06802E-05
0.0003396911
-5.222127E-05
0.0003490909
1.36333E-05
0.000351168
5.805985E-05
0.0003521771
6.799313E-05
0.0003479298
2.233027E-05
0.0003482742
5.639499E-05
0.0003481395
4.608237E-05
0.0003463104
-2.873936E-06
0.0003478921
7.507711E-05
0.0003443519
-0.0002333646
0.0007255496
0.001941235
The time was 00:00:00.5440568 for: iteration 0
Training: NaN
Training diverged. Aborting...
```

Let me turn KFAC back on, but with a smaller beta.

```
inl pars = {rate={weight=learning_rate; covariance=learning_rate ** 0.85f32; minibatch_size=size.minibatch}}
```

Whops, forgot about the covariance.

```
{rate = {covariance = 0.85; minibatch_size = 64; weight = 6.103516E-05}}
The learning rate is 2 ** -14
Starting timing for: iteration 0
The time was 00:00:07.1996102 for: iteration 0
Training: 2.42574334552046
The learning rate is 2 ** -14
Starting timing for: iteration 1
The time was 00:00:07.1170810 for: iteration 1
Training: 2.24516750402365
The learning rate is 2 ** -14
Starting timing for: iteration 2
The time was 00:00:07.1047007 for: iteration 2
Training: 1.99083418822299
The learning rate is 2 ** -14
Starting timing for: iteration 3
The time was 00:00:07.1160003 for: iteration 3
Training: 1.84775092602735
The learning rate is 2 ** -14
Starting timing for: iteration 4
The time was 00:00:07.1452626 for: iteration 4
Training: 1.79886565490982
```

Now that I've fixed covariance it works even with -14. It is not quite optimal, but it is close enough.

Interesting that I needed to divide by `k` rather than `minibatch_size`. `k` is number of timesteps times the minibatch size.

Let me take out the `minibatch_size` arguments from the two tests.

Done. Let me try running it with a default learning rate of zero.