``` {rate = {covariance = 0.85; weight = 6.103516E-05}} The learning rate is 2 ** -14 Starting timing for: iteration 0 The time was 00:00:07.0714271 for: iteration 0 Training: 2.43107565123971 The learning rate is 2 ** -14 Starting timing for: iteration 1 The time was 00:00:06.9877025 for: iteration 1 Training: 2.06160500030006 The learning rate is 2 ** -14 Starting timing for: iteration 2 The time was 00:00:06.9758889 for: iteration 2 Training: 1.97707993883694 The learning rate is 2 ** -14 Starting timing for: iteration 3 The time was 00:00:06.9942780 for: iteration 3 Training: 1.92417289940503 The learning rate is 2 ** -14 Starting timing for: iteration 4 The time was 00:00:07.0612741 for: iteration 4 Training: 1.89123285374851 ```
Yeah, it is not as good as a well tuned LR, but nonetheless if it means not having to actually tune it anymore then this is acceptable.

The real test though is how well it does on the poker game. If it does well, then I will make this my chosen method. If not, then I'll write off the work of the past few hours.

Let me give it a try.