``` The learning_rate is 2 ** -14 ------ The CudaRandom pseudorandom seed is 0 Starting timing for: iteration 0 CuSolve function failed! The leading minor of order 1 is not positive definite. C:/Users/Marko/Source/Repos/The Spiral Language/Temporary/bin/Release/cuda_kernels.cu:388: block: [0,0,0], thread: [0,0,0] Assertion `is_info_zero` failed.
Unhandled Exception: ManagedCuda.CudaException: ErrorAssert: A device-side assert triggered during kernel execution. The context cannot be used anymore, and must be destroyed. All existing device memory allocations from this context are invalid and must be reconstructed if the program is to continue using CUDA.
```

Not off to a good start.

Let me print out the lrs then.

```
1.377037E-08
-0.0001220269
1.377037E-08
-0.0001220269
NaN
NaN
```

Let me run this in the Powershell. I need a large buffer to see all the outputs.