https://arxiv.org/abs/1703.04782 Online Learning Rate Adaptation with Hypergradient Descent
Since layerwise hypergradient descent does not work on the RL game, let me put it on a separate branch just in case I need it later.

The next experiment will be to implement the global one. I do not think it will work given than the local version failed, but it needs to be done so I can put it behind me.