Yeah, it does work now. It is trully a great pity that hypergradient descent is a fluke as it would definitely be a great benefit to have some LR adaptation for RL tasks, but it cannot be helped.
That was a fairly intense marathon session following up from yesterday, so let me finally put my thoughts together.

The main point that needs to be highlighted - weight sampling actually gives a concrete benefit on the poker task. It is minor, but definite.

Given how rare definite gains are, I will gladly accept it. For the past six months I literally had only 3 gains of note: KFAC, going from PG to AC, and now weight sampling.

```
The learning_rate is 2 ** -13
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 233 and 767 out of 1000.
The time was 00:00:08.8129459 for: iteration 0
Starting timing for: iteration 1
Winrate is 388 and 612 out of 1000.
The time was 00:00:13.8791427 for: iteration 1
Starting timing for: iteration 2
Winrate is 597 and 403 out of 1000.
The time was 00:00:15.8891123 for: iteration 2
Starting timing for: iteration 3
Winrate is 649 and 351 out of 1000.
The time was 00:00:14.5758203 for: iteration 3
Starting timing for: iteration 4
Winrate is 680 and 320 out of 1000.
The time was 00:00:15.2811478 for: iteration 4
Starting timing for: iteration 5
Winrate is 696 and 304 out of 1000.
The time was 00:00:15.5226034 for: iteration 5
Starting timing for: iteration 6
Winrate is 695 and 305 out of 1000.
The time was 00:00:16.0762263 for: iteration 6
Starting timing for: iteration 7
Winrate is 740 and 260 out of 1000.
The time was 00:00:15.4605723 for: iteration 7
Starting timing for: iteration 8
Winrate is 733 and 267 out of 1000.
The time was 00:00:16.8748427 for: iteration 8
Starting timing for: iteration 9
Winrate is 733 and 267 out of 1000.
The time was 00:00:17.1975331 for: iteration 9
Starting timing for: iteration 10
Winrate is 750 and 250 out of 1000.
The time was 00:00:16.3701147 for: iteration 10
Starting timing for: iteration 11
Winrate is 746 and 254 out of 1000.
The time was 00:00:15.6491637 for: iteration 11
Starting timing for: iteration 12
Winrate is 724 and 276 out of 1000.
The time was 00:00:15.9862208 for: iteration 12
Starting timing for: iteration 13
Winrate is 714 and 286 out of 1000.
The time was 00:00:16.1337309 for: iteration 13
Starting timing for: iteration 14
Winrate is 741 and 259 out of 1000.
The time was 00:00:16.0180601 for: iteration 14
```

The test run with -13 finished. As expected even though I have an RNN as the input layer, the thing does perfectly in the end.

With this I am done. I've been thinking a lot of what I want to do last night.

What I really wanted to do is make a great player that can crush the weak players. Instead with this player here what I could do is let it engage in self play in preparation for the assault. Maybe after it plays a few million or billion of hands of NL Holdem it will learn something, maybe enough to beat the small stakes.

That is one thread I am going to follow. Though I want to go into probabilistic programming, it would be just too bad to not follow through with the work that I had done. I really want to create some work for the GPU at long last and have it train while I do some work on the side.

So that thread is to build out all the infrastructure and let that happen.

But fundamentally there is no way I could be satisfied with something like that. It would be different if the algorithm I have now was the pinnacle of RL, but I know it is just a piece of trash. Though KFAC + weight sampling is a definite improvement on model-free PPO, ultimately it is not that different from OpenAI's failed Dota attempt.

It is unsuitable. Poker is not the kind of game I can expect to carry through in real life with the style of AlphaGo. The main issue is the rake - I need to be able to win by a significant amount over the weak players to make this profitable.

Well, Libratus did managed to eke out 10bb over the pros, so it might turn out fine. But...

There are just too many mysteries that I could not make a dent in despite all my efforts.

Just putting all my effort into engineering and leaving it all to a crap algorithm is not something that befits me - it is not the skill nor the power that I seek. Nor will such a mediocre effort open the path to the Singularity. Engineering is necessary, but not sufficient.

Towards that end, I will play it a bit smarter and try to open a crack from a different direction. I've finally realized that to get good exploration, the benefits of the Bayesian approach are not optional, but mandatory. For the sake of reaching the true tomorrow, the Bayesian ideal of rationality will have to be carved into code and ignited. There is no other choice.

Let me post an update in that RL sub thread (as I've linked to it from the PL monthly one) and then I will take a break from programming for the day. I need to figure out exactly what I want to do when I get back to it.

First of all, I think I should work on extending the game so it is full NL HUSNG. In order to get a sense for what the players are doing, just like last time, I might need to build interfaces and such. I'd really like it if I could play against the players I am training. It will be a challenge to make GUIs in Spiral, but I should manage it somehow.

After that comes self play and checkpointing. And at that point I really have everything I need - as past that I'll have the player, the task will shift to figuring out how to interface with online poker rooms.

Also, I do not want to say this, but since time is starting to be of the essence, I should consider doing all of this in a language other than Spiral. Though Spiral is excellent training, there is no point in having it do literally everything.

If I could use KFAC + weight sampling from PyTorch and Python gives me easy GUI libraries, I should just use it instead. After I start work on probabilistic programming, I should consider building a library in Spiral, but I should also consider doing the experiments in something else that does not have such large compilation times.

Actually, I hate Python so I do not think I will be using, but a dynamic functional language (like Clojure) would be quite acceptable for me. The functionally pure type level programming has given me a taste for dynamic languages.

Before the last six months, that is before I really started doing RL in earnest and was essentially designing a programming language, I really imagined that my progression would be towards greater and greater models and algorithms, but now that I've made little progress towards better algorithms apart from KFAC + weight sampling, and my model is basically a vanilla RNN, I have much less reason to stick with the present course. I do no need differentiable programming if I am just going to use a standard RNN. That thing would be fine even if I did it by hand in C and called it from somewhere else.

Well, that depends on how good the situation is elsewhere. I've never really bothered investigating my rivals properly, and now that Pyro is here, I have an actual reason to do so.

So before, I resume programming I should investigate the state of languages and their libraries.

I guess for the next two weeks I'll be studying probabilistic programming then.