11am. https://www.reddit.com/r/reinforcementlearning/comments/9z2b0t/r_practical_bayesian_learning_of_neural_networks/ebmf4rt/
"Let me do a little update, since I've linked to this thread from the [PL monthly one](https://www.reddit.com/r/ProgrammingLanguages/comments/a25wrj/december_18_monthly_what_are_you_working_on_how/eawspsa/?st=jpkz4tvr&sh=9c71914c) and you might be interested.

Now that I am done with the ML library redesign, I am finally testing the noisy KFAC on the poker toy game and it does seem like there is a minor, but notable benefit to robustness in using it there. Since the actual gains over the last six months specific to RL have been much rarer than I'd hoped, I'll keep using this. KFAC, going from PG to AC, learnable eligibility traces, and now weight sampling are literally the only gains over the standard model-free PG setup that I've gained in the last six months. Only the first two are not minor. It is quite pathetic and I'd expected a lot more.

Note that since the effect is not large, I cannot tell how well this would do on larger and more varied tasks. Neither have I compared weight sampling to an alternative regularization scheme like the cross entropy bonus as I do not have it implemented. As those tests would be time consuming I do not have the intent nor the interest in doing them myself. Caveat emptor.

Also the intrinsic motivation idea is not something I will try for two reasons:

* I imagined that bootstrapping + weight sampling would measure the scale of the means of the value rather than the scale of the value, but after thinking about it more in depth, I realized there is no reason to think that would happen. Hence, my conclusion is that intrinsic motivation will require the full Bayesian inference. I just can't tell how to measure the scale of the means of the values with my current level of understanding.

* Actually using weight sampling in the final AC layer makes things worse. I only observe a benefit when using it in the RNN layer below it.

---

Also, I've spent over half a day implementing and testing out [hypergradient descent](https://arxiv.org/abs/1703.04782). With KFAC, it works well on the Mnist and char-RNN tests, though not as well as well tuned static learning rate. On the poker test that I've most hoped it would work, it just dies. It cannot go a step without blowing up no matter what I set the hyper learning rate to."

11:35am. https://www.reddit.com/r/reinforcementlearning/comments/a5gi3u/reinforcement_learning_predictions_2019/ebmg2vr/

"My internal predictions over the last six months have been fairly horrible which is making me thing that making real progress in RL will require giving it the Bayesian treatment. I've been impressed by some of the things I've seen in probabilistic programming, so it might worth looking for the next break there. I am throwing in the towel on the current strain of RL - it was only good for making me exercise my programming skills.

Right now my view is that it is not the deep nets that are the problem, but the way they are being optimized. There is a bunch of things which I do not understand at all, but has caught my attention like [inference compilation](https://youtu.be/0ePp_vFDZII) which allows one to do exact Bayesian inference using deep nets.

For a while now, I've been thinking on how to resolve the deadly triad situation in deep RL which occurs when bootstrapping, non-linear function approximation and off-policy training is combined. One thing that occurred to me regarding off-policy training is that feeding the inputs in arbitrary order to a deep net is really quite different from Bayesian conditioning on them. Having that thought in mind might turn up something."

Also the tests are finally done with both the small and large RNN. 1 run out of 10 is kind of bad. Maybe this is due to the learnable eligibility trace? I don't know.

I said that the benefit was notable, but now I am not sure. I thought that weight sampling would make it robust to random seeds, but that is not the case at all.

Aagh, that is deep RL in a nutshell for you. This is not real skill. This is not something I can rely on.

I can take pride in Spiral and the things I have created, but this algorithm will not get me far. Not like this. I am supposed to believe this can beat Dota when it can barely do a toy poker game consistently. Yeah, sure.

It is a failure. And then you lose in a game what you do is restart and do it better the next time around.

It is just bad luck. I did not want to believe that supervised learning gains do not translate straightforwardly to reinforcement learning, but they do not. I need an alternative vision of the future at this point.

Since I am pretty good at programming, before I do anything else, I think I should learn some statistics in the form of probabilistic programming. That will be a real skill.

Let me take a break here.