The chapter 3d of the Bayesian RL book is a bit weird. I am not sure that some of the examples are working properly. Forget that for now. I'll continue it tomorrow.
I am despondent, but ultimately the only thing that matter is what I want to do.

What I want to do is forget deep nets for a while and do a thorough study of probabilistic programming.

Over six months ago, when I was first starting the RL journey, my original plan was to use Q learning, but somehow I've been herded by circumstances into constantly compromising.

What I should do is pick Bayesian Q Learning as chosen method and then find a optimizer that is not donkey retarded and solve the problem that way - with a simple linear net and do so 100% of the time.

I do not care if I have to chose HMC or some other optimization scheme as long as it is Bayesian. After that I'll figure out how to bring in deep nets and more in that context. Deep nets do not matter, the basics do. Moving from tabular to deep RL has given me a blow to the robustness I haven't been able to recover from.

And if Bayesian inference is doing its job properly then it should not succumb to the deadly triad regardless of the architecture. The deadly triad only exists because of the crapiness of current optimizers and nothing else.

I do not care if this part takes me as much as 3 years - by the end of it I will absolutely master the statistics side of ML and thoroughly merge it with programming.