"1:05pm. Let me check out the uses of HMC with Q learning.
http://proceedings.mlr.press/v37/betancourt15.pdf
"Leveraging the coherent exploration of Hamiltonian flow, Hamiltonian Monte Carlo produces computationally efficient Monte Carlo estimators, even with respect to complex and high dimensional target distributions. When confronted with data-intensive applications, however, the algorithm may be too expensive to implement, leaving us to consider the utility of approximations such as data subsampling. In this paper I demonstrate how data subsampling fundamentally compromises the scalability of Hamiltonian Monte Carlo."

So much about HMC. I am also done with the videos. They are not that interesting apart from the few I picked originally. Stuart Russel's lecture was interesting. I might post it on the ML sub.

1:20pm. https://www.reddit.com/r/MachineLearning/comments/a5t7wu/d_stuart_russel_probabilistic_programming_and_ai/

```
Video (54m): https://youtu.be/JzBrp5LnNCo

This video is the most interesting of the bunch from the recent PROBPROG conference in my view. Here is a short transcript of a part where a shot was fired at deep learning. It is a juicy piece of meat to chew on.

https://youtu.be/JzBrp5LnNCo?t=655

---

...And so you are throwing away a massive amount of generalization capability if you do not have this kind of expressive power.

And I believe, though this is not a particularly well explored area of computational learning theory that, you could also think, not so much what is the expressive power of the language, but how fast is it to evaluate the function? If it is a linear time function, that your hypothesis applied to an example is linear time then that means that if you are trying to learn a function from a family that is NP-Hard, you are going to in many cases need an exponentially large representation for the function and therefore again you are going to need a vast number of examples.

So when someone says (dismissive wave of hand) neural nets are really fast, and you are trying to do inference with Bayes nets and that's really slow. I am not going to even talk to you because you can't run the Bayes net on the entire web at once...then don't feel bad. The fact that your Bayes net is taking a long time is a good sign. At least it has a chance of forming a concise representation of the function that you are trying to learn. And I think the same argument in spades extends to things that are much more expressive than Bayes nets. So I do not think it is a scale the more complex your language the larger the time complexity of inference the worse things get. In fact it may be that the better things get.

So I suspect that the deep learning community may be repeating or is already repeating historical mistake that happened in AI. This is Bad Old Fashioned AI (in the 1960s). This is preformal (audience laughter), this is before we even sort of understood how to do first order inference and all that stuff, but people showed "Wow, with GPS, General Problem Solver we could produce three, four step plans so it is just a matter of scaling up, right? You know, computers will deal with this, they're effectively infinitely fast."

This is before we had the theory of NP-Hardness and all that stuff, so they were completely ignoring computational intractability. They eventually called this the combinatorial explosion. And then they tried other ways of building AI systems. We are doing the same thing now except with sample intractability. We are looking at approaches to problems that are going to require an exponentially or super-exponentially many examples. The universe is just not big enough to contain that much data for some of these problems.
```

2pm. Let me check out the lecture on sequential MC. This is what I will probably be using.

https://arxiv.org/abs/1708.06040
Meta-Learning MCMC Proposals

I'll read this paper after that. I'll probably implementing this eventually, probably months from now. Right now, my first goal will be to understand the very basics of how PP languages are constructed.

http://bayesiandeeplearning.org/2016/papers/BDL_41.pdf
Nested Compiled Inference for Hierarchical Reinforcement Learning

Maybe this too.

https://www.youtube.com/watch?v=tNBibTnaJ6s
Lecture 22: Introduction to Sequential Monte Carlo Methods

Here is the lecture I am talking about."