Somehow I do not feel like reading the Bayesian RL book at all. It is starting to get complex which is not a problem per se, but I am craving to get some hands on experience at this point.
I want to just dive in.

Before I can do that, though I want to study Bayesian inference algorithms. I heard about a bunch and I do not understand any of them at this point.

My goal is very, very simple at this point. I just want to go real step beyond tabular methods and solve the one card poker task in a reliable manner.

Before I even start that, I guess I should start with Gridworld and Bandit problems. And forget the GPU, do it on the host for the time being.

What should the first thing I study be? Anglican, WebPPL?

http://dippl.org/
The Design and Implementation of Probabilistic Programming Languages

Maybe this? I think I'll start with this thing and Anglican.

But before that let me watch the lectures on those Bayesian inference methods that I linked to earlier. This will take me a day. After that I will look for papers on various methods.

Only after that will I study DIPPL and Anglican.

I'll go with that plan.

The main priority is to improve my understanding of this. There are gaping holes in my mind with regards to Bayesian inference and ProbMods and the Bayesian RL book is just the appetizer.

I'll study, and study, and study and eventually get through this.

I understand well enough by now, I only get rewarded once I do it properly.

That having said, I admit that there are some bonuses to the Bayesian approach apart from improved optimization.

Depending on how things go, the improved interactivity might mean that I might not necessarily have to let lose the agents directly on the unsuspecting populace. I could use the trained agents as my own trainers and analysts and use them to teach myself how to play.

To start things off, it might be easier to do it that way that engineer all the machinery to have them do automatically. And it would be great intro to all the art stuff this could potentially enable.

Those elite agents are the key to everything. Whatever I choose, having them in my hands will unlock so many options.

Ahhh, I am truly so thirsty and so very hungry. The goal seems so far away...

Yet I will never stop moving forward. If I could do it for 4 years, I could do it for 4 more.