"10am. Ok, let me start the talk. Forget the residual learning thing. I'll just believe that Bayesian inference will do the right thing.
10:05am. https://youtu.be/Z2DvWv1vh40?t=4353
"The optimal q is the posterior distribution."

Yeah, I understand all of it I think. The goal of Bayesian inference is to make the proposal distribution as close as possible to the posterior one.

After I am done with this I will go to the ML book and take a look at the toy example once more.

10:30am. I think residual learning will work once I take the likelihoods in mind. Only the naive algorithm is the problem. With importance sampling training on past data should not be even the slightest bit of the problem. Though to acount for shifting opponents I'll have to get rid of old data periodically.

All in all, I think that the move from frequentist ML to Bayesian one will be straight forward. I think that thanks to all the examples, I've internalized something very important. Now all I need to do is get some practice in it.

I really, really bothed it during the PGM course. I learned pretty much nothing there.

Damn, the product rule is so damn misleading. I can't believe I never realized the flow perspective of things and the link between the posterior and the proposal distribution. I really completely missed the point of the likelihood probability.

10:45am. https://youtu.be/Z2DvWv1vh40?t=5672

I'll have to take a look at what FIVO is.

10:55am. Done with the talk.

Going from deep learning to Bayesian VI, in the current library won't be as huge step as it might seems. I'll replace sampling from the curvature with sampling from the proposal distribution. I'll make sure to add costs based on the prior during that - this is one of the huge problems that I have currently which is propably the reason why I am getting no regularization benefit with noisy KFAC. I'll also get rid of all the useless costs and replace them with factor statements. So no more cross entropy and squared error nonsense. For squared error, I might need to do it in terms of observe which I suspect can be implemented with factor statements.

I am actually in pretty good condition here, but I need to figure out what the basic Monte Carlo techniques are doing. I'll start work towards mastering that after the break."