It is finally time that I do some programming. This next time will be like old times sake. Remember in 2016 when I was doing those shitty code generators that eventually became Spiral?
So it will be this time. It is time to go on another adventure.

To start things off, I will open another project folder. I'll leave the Learning library and the poker Games alone for a while.

The Probabilistic project will be used for Bayesian inference.

My first goal will be to get the enumeration inference over discrete variables done. Before that I will start by implementing sampling.

Rather than use straight out code transformation like WebPPL and other languages do, I will just use monadic code to do all the CPSing. That is how it should be done in functional languages. It should be quite a workable solution.

Finishing all of this is not something that will take my only a few days, but will be an ongoing effort over the following weeks, months and years.

And today is when it will all start.

Looking just a bit into the future, I think that Bayesian Q learning should be the first workable algorithm that will drop out of this work.

The AC currently sitting in the ML library is trash and I want to have nothing to do with it.

The Probabilistic project represents a new era for me - one where I get rid of useless, ad-hoc methods and replace them with sensible, rationally motivated ones.