"2:10pm. I am thinking.
2:20pm. I am thinking some more. It has been a while since I exercised these parts of my brain.

I think for the enumeration based methods, I should just collect the traces in an immutable map and use integers as keys. That would be the easiest way. It will allow for a straightforward implementation of both sample and factor. It will be a piece of cake.

Of course of sampling, I won't collect the values in a dictionary, but simply propagate them back.

I guess I will need some way of storing the proposal distributions.

3:30pm. Had to take a break. Honestly, I am still thinking about it. I really confused as to how to do variational inference. I mean is there any other choice but to just pass in the gradient at the final output?

In one video I watched, the guy did compare it to reinforcement learning. Otherwise how the hell would I even optimize that ELBO thing?

3:35pm. Ah, wait, I see. You could do a Bayesian AC easily enough.

```
factor (reward * (R - V))
```

Just write this thing and there it is.

I am a bit confused as to how I could normalize the factors though. Unlike with sampling there simply won't be anything to normalize with.

Hmmm...so maybe some form of sampling would be necessary if just for that?

I've been really scratching my head over how to do it.

3:40pm. I think I understand how to propagate the scores backwards with regular sampling and enumeration, but I admit I am a bit confused as to how to interleave that with sampling and factor statements.

Do I maybe not need to do it as it would be all reflected in the final log score? That could be it.

I think I could go with that.

3:50pm. No I got it wrong, it is not that simple. Remember those examples where I return an earlier statement? That is proof that I cannot just pass the gradient into the output.

So it is not at all like reinforcement learning. How troublesome.

Yeah, I think I am confused as to how to use the gradients after all. Should I just get enumeration out of the way?

Let me make a toy example.

Maybe I'll call it a day after that.

Actually, before that let me take one more look at the naive residual learning example in the Sutton and Barto book. I haven't though about it before, but what should its behavior be if it was weighted? I remember reading before that for residual learning to work, it needs alternative samples or other such nasty hacks. Maybe simply weighting it by the policy probability would work? It would be cool if that was the case."