I am just overthinking it. The issue is not about allowing gradients to flow through to the next step like in standard Q learning. Rather the issue is that I do not have any place to let them flow in the first place. It seemed easier at first glance, but I have nothing to optimize with in order to get closer to the posterior.
Before I do variational inference that kind of thing would be really important to figure out first. Just what is that KL divergence for the ELBO?

Just thinking about it will not get me closer to the answer - I never had that kind of brainpower. Instead I will have to pick it up from somewhere. For now let me do a bit of programming and then I will call it a day.