"9:35am. https://youtu.be/ogdv_6dbvVQ?t=1311
No seriously, how could it possible be possible to optimize proposal so it gets closer to the posterior when you have no idea what the posterior is?

"Hey how do you computer the KL divergence when it requires knowing the posterior itself?"

Oh, so he will finally get to it.

9:40am. Wait, could it really be that is doing something as stupid as sampling and then optimizing in the direction of the greater example?

9:45am. Damn, what are these two terms? If I cannot figure it out after the talk is done, I'll just ask and if that fails look into how it is done inside the PP libraries.

9:55am. https://youtu.be/ogdv_6dbvVQ?t=1989

What is natural gradient being talked about here?

10am. https://youtu.be/ogdv_6dbvVQ?t=2310

This might go somewhere.

https://youtu.be/ogdv_6dbvVQ?t=2346

So it is like this? The prior `p(z)` is optimized independently of the cost at the end? And the proposal distribution is negative in the cost?

Wait, there is something I am confused about. `q(z)` and `p(z)` here...what exactly is the difference between them?

10:05am. https://youtu.be/ogdv_6dbvVQ?t=2392

Ok, this is definitely interesting. I am not sure what exactly am I looking at here.

10:15am. https://youtu.be/ogdv_6dbvVQ?t=2669

Ok, so this is a bit like reinforcement learning. That having said, I still do not get how to take the derivative of the excuted model. It is not a cost function that I could optimize even if the probability `p(x,z)`, namely the forward pass is a proxy for it.

10:20am. https://youtu.be/ogdv_6dbvVQ?t=2755

This seems simple enough...

https://youtu.be/ogdv_6dbvVQ?t=2790

Ahhhhhhhhhhhhhhhhhhhhhhhhhhh...I get it. Agh, I get it. I really do get it now.

That gradient `grad log q(z,v)` is not the gradient of the function with respect to the output like in deep learning. It is literally just a sample from the proposal distribution `q`. Meaning it is the gradient of `m + s * e` while ignoring everything else.

I remarked earlier that Bayesian inference is kind of like a RL problem with a global reward...and this is in fact exactly what is going on here. `log p(x,z)` is literally never taken the derivative of in stochastic variational inference.

All this time I've been busting my head trying to figure out how VI works when it turns out there is nothing much to figure out at all.

This is not that great...if there would be a way around it that would be...to make a really large variational family - as large as the net it is trying to optimize.

10:35am. I admit I thought it would be a better idea than this. If it is like so then it is quite obvious why BNNs with mean field VI do not work as well as the regular ones. Global reward signal will simply not cut it here.

10:45am. I am getting lost in thought. Let me resume the talk."